requirements.txt
Django>=4.0,<4.3
djangorestframework>=3.14.0
djangorestframework-jsonapi>=6.0.0
django-filter>=23.1
drf-spectacular>=0.23.0
psycopg2-binary>=2.9.0
flake8>=5.0.0
